- made subprocess loging work on windows
